implement return resolution choosing option, and save memory in intermediate functions

+++ wfdb/readwrite/downloads.py
@@ -58,8 +58,7 @@ def streamdat(filename, pbdir, fmt, bytecount, startbyte, datatypes):
     sigbytes = r.content
 
     # Convert to numpy array
-    # Cast to int64 for further processing
-    sigbytes = np.fromstring(sigbytes, dtype = np.dtype(datatypes[fmt])).astype('int')
+    sigbytes = np.fromstring(sigbytes, dtype = np.dtype(datatypes[fmt]))
 
     return sigbytes
